+++ integreat_cms/cms/views/analytics/translation_coverage_view.py
@@ -33,9 +33,11 @@ class TranslationCoverageView(TemplateView):
         translation_coverage_data = {}
         outdated_word_count = Counter()
 
-        pages = region.pages.filter(explicitly_archived=False).cache_tree(
-            archived=False
-        )[0]
+        pages = (
+            region.pages.filter(explicitly_archived=False)
+            .prefetch_major_public_translations()
+            .cache_tree(archived=False)[0]
+        )
         for language in region.active_languages:
             language_coverage_data = Counter()
             for page in pages: