+++ integreat_cms/cms/views/pages/page_tree_view.py
@@ -97,7 +97,9 @@ class PageTreeView(TemplateView, PageContextMixin):
             page_queryset = region.pages.all()
         else:
             page_queryset = region.pages.filter(lft=1)
-        pages = page_queryset.cache_tree(archived=self.archived)[0]
+        pages = page_queryset.prefetch_major_public_translations().cache_tree(
+            archived=self.archived
+        )[0]
 
         if filter_data:
             # Set data for filter form rendering